---- Version 1.3 StarShooter ---- + Better design. Shooting lasers are now added and linked to the chosen team. Explosion of ships are now fixed. + Score indications. When winning, the popup view now displays the score as the number of shots fired to win. + Bug correction : When winning, the view reload and the game restart even if no parameters are modified.
This version allow you to play basically to the game.
You can choose a number of ships, a team, a ship for your team.
When the game start, just one ship is different meanings that it's the intruder you have to touch on to win.
Sounds and vibrations happen when touching on the screen.
When winning, the ship change to an explosion and a popup window shows you that you win to finally get you back on the menu.

1.2 : Preferences are now saved when application's killed.
 master